move event status and on schedule every fields into plan package
